refactored tests to dedicated files
