+++ com.unity.render-pipelines.high-definition/Tests/Editor/ArrayList.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1e8f2c56a2802734596005d9e3605c7f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
