+++ com.unity.render-pipelines.high-definition/Tests/Editor/ArrayList/ArrayListTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3bff387107ac02249bb0eeeea6bac94a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
