+++ com.unity.render-pipelines.high-definition/Tests/Editor/Memory.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 12532ed62ccd79f4496245b6f0e3c9cf
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
