+++ com.unity.render-pipelines.high-definition/Tests/Editor/Memory/FixedAllocatorTests.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 6c671e3d22efb7a4dbec943fc3a5285f
+guid: 8276f95304cbc9649a45f3f65152bb97
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
